shop/requirements.txt
# commented for now cause will be cleared before prod
django==3.2.3
# django-extensions==3.1.3
ipython==7.27.0
flake8==3.9.2
flake8-django==1.1.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
# faker==8.5.0
# pydot==1.4.2
# django-debug-toolbar==3.2.1
# django-silk==4.1.0
# django-redis==5.0.0
# redis==3.5.3
# better-profanity==0.7.0
django-celery-results==2.2.0
# djangorestframework==3.12.4
# markdown==3.3.4
# django-filter==2.4.0
requests==2.26.0
django-bulk-update-or-create==0.3.0
